// apps/web/src/proto/LeaveReason.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type LeaveReason = { "kind": "quit" } | { "kind": "timeout" } | { "kind": "kicked", by: string, reason: string | null, } | { "kind": "banned", by: string, reason: string | null, until: number | null, } | { "kind": "server_shutdown" } | { "kind": "replaced" };