// apps/web/src/proto/Welcome.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Channel } from "./Channel";
import type { Client } from "./Client";
import type { Group } from "./Group";
import type { IceServer } from "./IceServer";
import type { Member } from "./Member";
import type { Permission } from "./Permission";
import type { ServerInfo } from "./ServerInfo";
import type { Unread } from "./Unread";

export type Welcome = { session: number, uid: string, server: ServerInfo, permissions: Array<Permission>, groups: Array<Group>, channels: Array<Channel>, clients: Array<Client>, 
/**
 * Known users of this server (online ones are also in `clients`).
 */
members: Array<Member>, 
/**
 * Read state of every channel this user can read.
 */
unread: Array<Unread>, ice_servers: Array<IceServer>, };